+++ src/SmartOffice/Models/Role.cs
@@ -1,0 +1,21 @@
+﻿// -----------------------------------------------------------------------
+// <copyright file="GraphProvider.cs" company="Microsoft">
+//     Copyright (c) Microsoft Corporation. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------
+
+namespace Microsoft.Partner.SmartOffice.Models
+{
+    public class Role
+    {
+        /// <summary>
+        /// Gets or sets the description for the role.
+        /// </summary>
+        public string Description { get; set; }
+
+        /// <summary>
+        /// Gets or sets the name associated with the role.
+        /// </summary>
+        public string DisplayName { get; set; }
+    }
+}